refactor(config): Rename settings to config
