+++ lib/src/models/config.dart
